requirements.txt
absl-py==2.1.0
ale-py==0.10.2
annotated-types==0.7.0
AutoROM==0.4.2
AutoROM.accept-rom-license==0.6.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorlog==6.9.0
contourpy==1.3.1
cycler==0.12.1
decorator==5.2.1
docker-pycreds==0.4.0
docstring_parser==0.16
etils==1.12.1
Farama-Notifications==0.0.4
filelock==3.17.0
fonttools==4.56.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
glfw==2.8.0
grpcio==1.70.0
gymnasium==1.1.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.8
loguru==0.7.3
lz4==4.4.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
moviepy==2.1.2
mpmath==1.3.0
mujoco==3.3.0
networkx==3.4.2
numpy==2.2.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==10.4.0
platformdirs==4.3.6
prettytable==3.15.1
proglog==0.1.10
protobuf==5.29.3
psutil==7.0.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyOpenGL==3.1.9
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
seaborn==0.13.2
sentry-sdk==2.22.0
setproctitle==1.3.5
Shimmy==0.2.1
shtab==1.7.1
six==1.17.0
smmap==5.0.2
stable_baselines3==2.6.0
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
toml==0.10.2
torch==2.6.0
torch-tb-profiler==0.4.3
tqdm==4.66.6
triton==3.2.0
typeguard==4.4.2
typing_extensions==4.12.2
tyro==0.9.17
tzdata==2025.1
urllib3==2.3.0
wandb==0.19.8
wcwidth==0.2.13
Werkzeug==3.1.3
zipp==3.21.0
